+++ plugin/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaRouteExplorerPanel.kt
@@ -74,6 +74,7 @@ class ArmeriaRouteExplorerPanel(
                 }
             })
             add(ArmeriaGenerateHttpRequestAction { selectedRouteFromTree() })
+            add(ArmeriaSyncRuntimeRoutesAction())
         }
         toolbar = ActionManager.getInstance().createActionToolbar("ArmeriaRouteExplorer", actionGroup, true).also {
             it.targetComponent = this